fix(step1) : step1 패키지 명 수정
